plots misha 1D tracks and annotations

shaman_gplot_tracks_and_annotations

shaman_plot_tracks_and_annotations(
  genome,
  interval_range,
  misha_tracks = list(),
  mt_colors = getOption("shaman.track_colors"),
  mt_ylims = NULL,
  annotations = list(),
  a_colors = getOption("shaman.annotation_colors"),
  add_genes = T,
  add_ideogram = T,
  add_axis = T,
  gene_stacking = "squish",
  gene_size = 0.7,
  track_size = 0.8,
  annotation_size = 0.7
)

Arguments

genome

Name of reference genome (e.g. "hg19", "mm10")

interval_range

1D interval (chrom, start, end) specifying the region to plot

misha_tracks

List of 1D track expressions (virtual tracks also supported) which can be extracted from.

mt_colors

Array of colors, one for each misha_track, which will be used to plot each 1d track.

annotations

List of gintervals highlighting annotated regions.

a_colors

Array of colors, one for each annotation set.

add_genes

Boolean flag indicating whether to show gene annotations.

add_ideogram

Boolean flag indicating whether to add an chromosomal ideogram.

add_axis

Boolean flag indicating whether to show chromosmal axis.

gene_stacking

Describes the viewing option of the genes track. Can be either "squish", or "dense"

gene_size

Size of gene annotation view

track_size

Size of track view

my_ylims

Y-axis limits for each misha track. If not provided, generated automatically based on the data to be displayed in the region.

annot_size

Size of annotation view

Details

Plots 1D tracks in UCSC format. Along with gene annotations and chromosomal axis, this function plots one dimensional misha tracks (e.g. rna-seq, chip-seq data) and marks annotated intervals in the plotting region.

Examples

gsetroot(shaman_get_test_track_db())
#> Error in setwd(groot): cannot change working directory
shaman_plot_tracks_and_annotations("hg19", gintervals(2, 175.5e06, 177.5e06))
#> Error: Invalid interval (chr2, 1.755e+08, 1.775e+08): end coordinate exceeds chromosome boundaries