generate a score matrix for observed data based on the expected

shaman_score_hic_mat_for_track

shaman_score_hic_mat_for_track(track_db, work_dir, obs_track_nms, exp_track_nms,
  points_track_nms, chrom, start1, end1, start2, end2, expand = 2e+06,
  k = 100, min_dist = 1024)

Arguments

track_db
Directory of the misha database.

work_dir
Centralized directory to store temporary files.

obs_track_nms
Names of observed 2D genomic tracks for the hic data. Pooling of multiple observed tracks is supported.

exp_track_nms
Names of expected (shuffled) 2D genomic tracks. Pooling of multiple expected tracks is supported.

points_track_nms
Names of 2D genomic tracks that contain points on which to compute normalized score. Pooling points from multiple tracks is supported.

chrom
The chormosome of the matrix.

start1
The start coordinate of the first dimension.

end1
The end coordinate of the first dimension.

start2
The start coordinate of the second dimension.

end2
The end coordinate of the second dimension.

expand
Size of expansion, points to include outside the matrix for accurate computing of the score. Note that for each observed point, its k-nearest neighbors must be included in the expanded matrix.

k
The number of neighbor distances used for the score. For higher resolution maps, increase k. For lower resolution maps, decrease k.

min_dist
The minimum distance between points.

Details

This function extracts observed data and expected data in an expanded matrix and computes The score for a point is the KS D-statistic of the distances to the points k-nearest-neighbors